requirements.txt
pandas
numpy
selenium
transformers
torch
sentence-transformers
matplotlib
jinja2
python-dotenv